Makes API hook
