+++ src/hooks/apiHook.jsx
@@ -1,0 +1,31 @@
+import { useState, useEffect } from "react";
+
+export default function useApi(url) {
+  const [data, setData] = useState([]);
+  const [isLoading, setIsLoading] = useState(false);
+  const [isError, setIsError] = useState(false);
+
+  useEffect(() => {
+    async function getData() {
+      try {
+        setIsLoading(true);
+        setIsError(false);
+        const fetchedData = await fetch(url);
+        const json = await fetchedData.json();
+
+        if (fetchedData.ok) {
+          setData(json);
+        } else {
+          console.log(json.errors);
+          throw new Error();
+        }
+      } catch (error) {
+        setIsError(true);
+      } finally {
+        setIsLoading(false);
+      }
+    }
+    getData();
+  }, [url]);
+  return { data, isLoading, isError };
+}
