+++ src/pages/Product/index.jsx
@@ -1,36 +1,14 @@
 import { useParams } from "react-router-dom";
-import { useState, useEffect } from "react";
+import useApi from "../../hooks/apiHook";
 
 export default function Product() {
-  const [data, setData] = useState(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [isError, setIsError] = useState(false);
   let { id } = useParams();
-
-  useEffect(() => {
-    async function getData(url) {
-      try {
-        setIsLoading(true);
-        setIsError(false);
-
-        const res = await fetch(url);
-        const json = await res.json();
-
-        setData(json);
-      } catch (e) {
-        console.log(e);
-      } finally {
-        setIsLoading(false);
-      }
-    }
-
-    getData(`https://api.noroff.dev/api/v1/online-shop/${id}`);
-  }, [id]);
+  const url = `https://api.noroff.dev/api/v1/online-shop/${id}`;
+  const { data, isLoading, isError } = useApi(url);
 
   if (isLoading || !data) {
     return <div>Loading Data</div>;
   }
-
   if (isError) {
     return <div>Error</div>;
   }
